requirements.txt
accelerate==1.6.0
certifi==2025.4.26
charset-normalizer==3.4.2
et_xmlfile==2.0.0
filelock==3.18.0
fsspec==2025.3.2
hf-xet==1.1.1
huggingface-hub==0.31.1
idna==3.10
Jinja2==3.1.6
joblib==1.5.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==1.24.3
openpyxl==3.1.5
packaging==25.0
pandas==2.2.2
psutil==7.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.13.3
torch==2.7.0
tqdm==4.67.1
transformers==4.31.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
